fix: correct .vue import paths after editor/game/wizard split
The move script matched extensionless imports but not '*.vue' ones, so
MarbleEditor.vue still imported the moved components from the old root paths and
EditorPalette pointed one level too high for its same-folder icon. Vite failed
to resolve them at serve time even though vue-tsc did not. All relative imports
now resolve and the app loads.
