+++ src/views/Games/MarbleEditor/MarbleEditor.vue
@@ -31,10 +31,10 @@ import { EDITOR_MENU_MAPPING } from './config'
 import { useMarbleEditor } from './editor/useMarbleEditor'
 import { useMarbleRace } from './game/useMarbleRace'
 import { useMarbleEditorSession } from './useMarbleEditorSession'
-import EditorPalette from './EditorPalette.vue'
-import EditorToolbar from './EditorToolbar.vue'
-import MarbleEditorLobby from './MarbleEditorLobby.vue'
-import MarbleEditorSummary from './MarbleEditorSummary.vue'
+import EditorPalette from './editor/EditorPalette.vue'
+import EditorToolbar from './editor/EditorToolbar.vue'
+import MarbleEditorLobby from './wizard/MarbleEditorLobby.vue'
+import MarbleEditorSummary from './game/MarbleEditorSummary.vue'
 import { SAMPLE_MAPS } from './sampleMaps'
 import { MARBLE_OPTIONS, DEFAULT_MARBLE } from '../MarbleMadness/config'
 import type { CameraMode, MePhase } from './types'
