+++ src/views/Games/MarbleEditor/editor/EditorPalette.vue
@@ -3,7 +3,7 @@ import { computed } from 'vue'
 import { Trash2 } from 'lucide-vue-next'
 import { LobbyUIButton, LobbyUIIconButton } from '@/components/LobbyUI'
 import { PIECE_CATALOG, PLACEABLE_PIECE_TYPES } from '../pieceCatalog'
-import PiecePreviewIcon from '../PiecePreviewIcon.vue'
+import PiecePreviewIcon from './PiecePreviewIcon.vue'
 import type { PlacedPiece, TrackPieceType } from '../types'
 
 const props = defineProps<{
